@@ -2,10 +2,10 @@
   "name": "api-besmart",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
-  
+  "main": "index.ts",
+  "type": "module",
   "scripts": {
-    "start": "node index.ts",
+    "start": "nodemon --exec node --no-warnings=ExperimentalWarning --loader ts-node/esm ./index.ts",
     "test": "Error"
   },
   "keywords": [],
@@ -13,18 +13,26 @@
   "license": "ISC",
   "devDependencies": {
     "@types/node": "^20.11.17",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.5"
   },
   "dependencies": {
+    "@types/axios": "^0.14.0",
+    "@types/express": "^4.17.21",
+    "@types/morgan": "^1.9.9",
+    "axios": "^1.6.8",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "express": "^4.18.2",
+    "esm": "^3.2.25",
+    "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.1.2",
+    "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
+    "node": "^22.2.0",
     "nodemailer": "^6.9.9",
-    "nodemon": "^3.0.3",
+    "nodemon": "^3.1.0",
+    "ts-node": "^10.9.2",
     "tsx": "^4.7.1"
   }
 }
